==> 04-create-attributes-and-ARM.ipynb <==
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.sql.types import ArrayType, IntegerType, FloatType, StringType

spark = SparkSession \
        .builder \
        .appName("frequent_itemsets") \
        .getOrCreate()

data = spark.read.parquet("../data/data_with_topics.parquet").dropna()

data.show(5)

## Convert topic probabilities and LIWC scores into "attributes" for Association Rules Mining. 

These attributes correspond to either the presence or absence of a given topic or emotional/moral language, based on thresholds. 

Eg: "document x has: topic 1, topic 4, moral language and negative emotional language"

#Create dictionary of topics where each topic has a tuple as value: 

#First value of tuple corresponds to threshold to consider a topic probable enough
#in a document (0.20 for all topics, decided after observing a distribution of
#topics probabilities and manually checking for results of assignments.

#Second value is just the number of the topic to identify the document

mapping_topics = {
    'topic0': (0.2, 0),
    'topic1': (0.2, 1),
    'topic2': (0.2, 2),
    'topic3': (0.2, 3),
    'topic4': (0.2, 4),
    'topic5': (0.2, 5),
    'topic6': (0.2, 6),
    'topic7': (0.2, 7),
    'topic8': (0.2, 8)}

#Same logic for dictionary construction of morality values

#Thresholds defined as higher than the means of scores in blogs for the development
#of the dictionary 
#Boyd, R. L. (n.d.). The Development and Psychometric Properties of LIWC-22.
mapping_emotions = {
    'emo_pos': (1.17, 11),
    'emo_neg': (0.81, 12),
    'moral': (0.28, 13)}

def create_items(input_df, mapping, output_col):
    # Create a new column with the integers corresponding to attributes where
        #values are higher than the threshold. 
        df = input_df.withColumn(
        "relevant",
        F.array(*[
        F.when(F.col(col) > threshold, F.lit(value)).otherwise(None) 
        for col, (threshold, value) in mapping.items()]))

        # Filter out None values from the array
        df = df.withColumn(f"{output_col}", F.expr("filter(relevant, x -> x is not null)"))
        df = df.drop("relevant")
        
        return df 

df_topics = create_items(data, mapping_topics, "relevant_topics")
df_liwc = create_items(df_topics, mapping_emotions, "relevant_liwc_scores")
df_topics.select('relevant_topics').show(5)
df_liwc.select('relevant_topics', 'relevant_liwc_scores').show(5)

df = df_liwc.withColumn("relevant_attr", F.concat(df_liwc["relevant_topics"], df_liwc["relevant_liwc_scores"]))
df.select('relevant_attr').show(5)

#Save file for usage later 
df.write.mode("overwrite").parquet("../data/data_with_attributes.parquet")

## Association Rules Mining

from pyspark.ml.fpm import FPGrowth

fp = FPGrowth(minConfidence=0.5, minSupport=0.1)
fpm = fp.fit(df.select(df.relevant_attr.alias('items')))
fpm.associationRules.orderBy("lift", "confidence", ascending=False).show(truncate=False)

- When family and mental health is discussed, it's 32% more likely to be discussed with negative emotional language
- When confidence and socializing is discussed, it's 28% more likely be discussed with positive emotional language
- When choices and struggles are discussed, it's 18% more likely to be discussed with negative emotional language
- When morality language is present, it is 16% more likely to be discussed with negative emotional language

fp = FPGrowth(minConfidence=0.5, minSupport=0.01)
fpm = fp.fit(df.select(df.relevant_attr.alias('items')))
fpm.associationRules.orderBy("lift", "confidence", ascending=False).show(truncate=False)

When choices, struggles and mental health are discussed with a moral tone, it's 56% more likely to be discussed with an emotional negative tone. 

fp = FPGrowth(minConfidence=0.5, minSupport=0.001)
fpm = fp.fit(df.select(df.relevant_attr.alias('items')))
fpm.associationRules.orderBy("lift", "confidence", ascending=False).show(truncate=False)

When choices and struggles + mental health + mindset and beliefs take a negative tone, they're 87% more likely to also have a moral tone. With a low support of 0.002, this translates to approximately 1,000 times this occured

### Create a table

import pandas as pd

d = {'antecedent': ["morality", "choices and strugges + family and mental health + moral tone",
                    "choices and strugges + family and mental health + mindset and beliefs + negative tone"] , 
     'consequent': ["negative emotion", "negative emotion", "moral tone"], 
     'lift': [1.16, 1.56, 1.87],
     'confidence': [0.51, 0.68, 0.51],
     'support': [0.14, 0.013, 0.002]}

df = pd.DataFrame(data=d)

with pd.option_context('display.max_colwidth', None):
    display(df)
