==> 08-community_detection.ipynb <==
import networkx as nx
import pandas as pd
import glob

# Load nodes
nodes_df = pd.read_csv("data_selfimprovement/network-analysis/nodes-topics-morality-net.csv")
nodes = nodes_df["node"].tolist()  # Convert to a list of nodes

# Load edges
csv_files = glob.glob("data_selfimprovement/network-analysis/edges_topics_morality_net/*.csv")

dfs = [pd.read_csv(file) for file in csv_files]

edges_df = pd.concat(dfs, ignore_index=True)
edges_df = edges_df[edges_df.weight >= 15]

edges_df.sort_values(by='weight', ascending=False)

# Create graph
G = nx.Graph()

#Add nodes
G.add_nodes_from(nodes)  

#Add edges
for _, row in edges_df.iterrows():
    G.add_edge(row['node1_norm'], row['node2_norm'], weight=row['weight'])

Basic exploration

nx.number_of_nodes(G)

nx.number_of_edges(G)

nx.density(G)

nx.number_connected_components(G)

# Get all connected components 
components = list(nx.connected_components(G))

# Find the largest component
largest_component = max(components, key=len)

# Subgraph of the largest component
G_red = G.subgraph(largest_component)

nx.number_of_nodes(G_red)

nx.number_of_edges(G_red)

nx.density(G_red)

Now, run community detection algorithm

communities = nx.community.louvain_communities(G_red, seed=123)

len(communities)

for community in communities:
    print(community)
    print("------------------------------")

{'criminal', 'fraud', 'country', 'high', 'greedy', 'house', 'disgrace', 'company', 'job', 'ungodly', 'thief', 'lewd', 'class', 'decently', 'move', 'chastise', 'duty', 'community', 'trouble', 'online', 'business', 'degree', 'innocence', 'university', 'misbehaving', 'ethic', 'irresponsible', 'molested', 'career', 'useless', 'parent', 'careless', 'ridiculously', 'blamed', 'child', 'disapproving', 'home', 'college', 'elitism', 'suck', 'movie', 'student', 'family', 'waste', 'money', 'lazy', 'corrupting', 'school', 'mother', 'horrible', 'car', 'pitiful', 'unqualified', 'sick', 'miser', 'kid', 'mom', 'corrupted', 'decent', 'scandal', 'snobs', 'disapproved', 'disapproves', 'study'}



{'sinner', 'generosity', 'process', 'cruel', 'foul', 'emotional', 'pressure', 'proud', 'ridicule', 'ideals', 'morality', 'cruelty', 'bravely', 'deceive', 'disapprove', 'emotion', 'sins', 'honour', 'truthfully', 'virtuous', 'truth', 'wronged', 'depression', 'ignorant', 'sinners', 'wrongly', 'vanity', 'unjust', 'selfish', 'journey', 'anger', 'ethical', 'pettiness', 'future', 'honor', 'honoring', 'sincere', 'sinister', 'heroic', 'sincerity', 'phony', 'deserve', 'sinned', 'value', 'inequity', 'desire', 'justice', 'choice', 'fault', 'behavior', 'unsavory', 'faults', 'deserving', 'courageous', 'patriot', 'commend', 'trustworthy', 'nobody', 'power', 'ignoble', 'worthwhile', 'generous', 'injustice', 'iniquity', 'character', 'present', 'mindset', 'absurd', 'acknowledge', 'equitable', 'grandiose', 'god', 'deserved', 'innocently', 'belief', 'mistakes', 'accept', 'betrayal', 'honorable', 'purpose', 'fairness', 'opinion', 'betrayer', 'truthful', 'scruples', 'corruption', 'ruthless', 'sinful', 'corrupt', 'honored', 'shame', 'unethical', 'worth', 'meaningful', 'mind', 'admit', 'courage', 'voice', 'therapist', 'betraying', 'regret', 'noble', 'brave', 'anxiety', 'challenge', 'unreasonable', 'reality', 'worthless', 'happiness', 'unfair', 'moment', 'society', 'liar', 'misuses', 'fear', 'angry', 'wrong', 'human', 'judge', 'heroes', 'immoral', 'prove', 'admirable', 'vengeance', 'adultery', 'heroines', 'dishonesty', 'therapy', 'feeling', 'prejudice', 'forgiving', 'boundaries', 'righteous', 'greed', 'virtue', 'wrongdoing', 'betrayed', 'unworthy', 'lie', 'unselfish', 'worthy', 'hypocrisy', 'ideal', 'excuse', 'hypocrites', 'forgiven', 'dignity', 'honest', 'evil', 'happy', 'abuse', 'sin', 'trust', 'true', 'failure', 'responsibility', 'tough', 'frauds', 'hero', 'selflessness', 'act', 'blame', 'cruelties', 'face', 'humane', 'treacherous', 'decision', 'forgive', 'sad', 'anxious', 'moral', 'absurdity', 'response', 'pain', 'honesty', 'forgives', 'success', 'conscience', 'comfort', 'environment'}



{'immodest', 'fakes', 'nefarious', 'petty', 'ridiculous', 'sexism', 'pedophile', 'toxic', 'outlaw', 'hair', 'arrogant', 'mistreating', 'hideously', 'nerd', 'offensive', 'nerdy', 'shamed', 'media', 'deceived', 'mistreat', 'dishonest', 'sissies', 'deceptive', 'accusation', 'haters', 'man', 'absurdities', 'confident', 'revolting', 'hypocrite', 'sneaky', 'account', 'spineless', 'pervy', 'disrespecting', 'partner', 'misogynistic', 'male', 'unacceptable', 'pussies', 'psycho', 'fun', 'guy', 'disrespectful', 'dumb', 'fake', 'sex', 'confidence', 'relationship', 'call', 'hideous', 'horrid', 'opinionated', 'comfortable', 'mistreated', 'equality', 'excusing', 'misconduct', 'depraved', 'judgy', 'personality', 'innocent', 'despicable', 'sexist', 'infidelity', 'puny', 'bigotry', 'inhumane', 'club', 'vile', 'vain', 'shit', 'kinda', 'fuck', 'dissing', 'perverted', 'pariah', 'overconfident', 'diss', 'dignified', 'shaming', 'accusations', 'hateful', 'reprehensible', 'weird', 'sissy', 'scum', 'cheat', 'faithful', 'inappropriate', 'conversation', 'overbearing', 'loyal', 'snide', 'bigoted', 'social', 'racist', 'untrustworthy', 'admonish', 'woman', 'unfaithful', 'instagram', 'excused', 'meet', 'group', 'prude', 'nice', 'slutty', 'interest', 'amoral', 'stupid', 'friend', 'pervert', 'betrays', 'cheats', 'disrespected', 'smug', 'misogynist', 'slandering', 'accuse', 'heroine', 'disrespect', 'standard', 'bigot', 'perverts', 'decency', 'respect', 'repulsive', 'betray', 'date', 'girl', 'cheating', 'treat', 'nerds', 'inconsiderate', 'porn', 'faking', 'attention', 'dude', 'outrageous', 'lazier', 'elitist', 'talk', 'molesting', 'slut', 'conceited', 'promiscuous', 'slander', 'traitor', 'attractive', 'rapist', 'promiscuity'}



{'project', 'energy', 'braver', 'dishonorable', 'misused', 'decadent', 'delinquent', 'play', 'food', 'conniving', 'drink', 'sleep', 'knowledgeable', 'healthy', 'useful', 'cruelly', 'lawlessness', 'youtube', 'infallible', 'lousy', 'information', 'productive', 'wicked', 'activity', 'learn', 'competence', 'read', 'specific', 'fatass', 'eat', 'development', 'task', 'water', 'teach', 'misuse', 'carelessness', 'doctor', 'hobbies', 'body', 'drug', 'routine', 'gluttony', 'cold', 'vilify', 'addiction', 'laziest', 'area', 'exercise', 'gym', 'book', 'hobby', 'humanitarian', 'laziness', 'smoke', 'skill', 'tv', 'wake', 'snob', 'music', 'competent', 'personal', 'list', 'content', 'video', 'diet', 'weight', 'internet', 'bed', 'health', 'practice', 'game', 'language', 'sneakily', 'sport'}




### Create table

d = {"Cluster 1": ["criminal, fraud, country, high, greedy, house, disgrace, company, job, ungodly, thief, lewd, class, decently, move, chastise, duty, community, trouble, online, business, degree, innocence, university, misbehaving, ethic, irresponsible, molested, career, useless, parent, careless, ridiculously, blamed, child, disapproving, home, college, elitism, suck, movie, student, family, waste, money, lazy, corrupting, school, mother, horrible, car, pitiful, unqualified, sick, miser, kid, mom, corrupted, decent, scandal, snobs, disapproved, disapproves, study"],
     
     "Cluster 2": ["sinner, generosity, process, cruel, foul, emotional, pressure, proud, ridicule, ideals, morality, cruelty, bravely, deceive, disapprove, emotion, sins, honour, truthfully, virtuous, truth, wronged, depression, ignorant, sinners, wrongly, vanity, unjust, selfish, journey, anger, ethical, pettiness, future, honor, honoring, sincere, sinister, heroic, sincerity, phony, deserve, sinned, value, inequity, desire, justice, choice, fault, behavior, unsavory, faults, deserving, courageous, patriot, commend, trustworthy, nobody, power, ignoble, worthwhile, generous, injustice, iniquity, character, present, mindset, absurd, acknowledge, equitable, grandiose, god, deserved, innocently, belief, mistakes, accept, betrayal, honorable, purpose, fairness, opinion, betrayer, truthful, scruples, corruption, ruthless, sinful, corrupt, honored, shame, unethical, worth, meaningful, mind, admit, courage, voice, therapist, betraying, regret, noble, brave, anxiety, challenge, unreasonable, reality, worthless, happiness, unfair, moment, society, liar, misuses, fear, angry, wrong, human, judge, heroes, immoral, prove, admirable, vengeance, adultery, heroines, dishonesty, therapy, feeling, prejudice, forgiving, boundaries, righteous, greed, virtue, wrongdoing, betrayed, unworthy, lie, unselfish, worthy, hypocrisy, ideal, excuse, hypocrites, forgiven, dignity, honest, evil, happy, abuse, sin, trust, true, failure, responsibility, tough, frauds, hero, selflessness, act, blame, cruelties, face, humane, treacherous, decision, forgive, sad, anxious, moral, absurdity, response, pain, honesty, forgives, success, conscience, comfort, environment"]}
     
d2 = {"Cluster 3": ["immodest, fakes, nefarious, petty, ridiculous, sexism, pedophile, toxic, outlaw, hair, arrogant, mistreating, hideously, nerd, offensive, nerdy, shamed, media, deceived, mistreat, dishonest, sissies, deceptive, accusation, haters, man, absurdities, confident, revolting, hypocrite, sneaky, account, spineless, pervy, disrespecting, partner, misogynistic, male, unacceptable, pussies, psycho, fun, guy, disrespectful, dumb, fake, sex, confidence, relationship, call, hideous, horrid, opinionated, comfortable, mistreated, equality, excusing, misconduct, depraved, judgy, personality, innocent, despicable, sexist, infidelity, puny, bigotry, inhumane, club, vile, vain, shit, kinda, fuck, dissing, perverted, pariah, overconfident, diss, dignified, shaming, accusations, hateful, reprehensible, weird, sissy, scum, cheat, faithful, inappropriate, conversation, overbearing, loyal, snide, bigoted, social, racist, untrustworthy, admonish, woman, unfaithful, instagram, excused, meet, group, prude, nice, slutty, interest, amoral, stupid, friend, pervert, betrays, cheats, disrespected, smug, misogynist, slandering, accuse, heroine, disrespect, standard, bigot, perverts, decency, respect, repulsive, betray, date, girl, cheating, treat, nerds, inconsiderate, porn, faking, attention, dude, outrageous, lazier, elitist, talk, molesting, slut, conceited, promiscuous, slander, traitor, attractive, rapist, promiscuity"],
      
    "Cluster 4": ["project, energy, braver, dishonorable, misused, decadent, delinquent, play, food, conniving, drink, sleep, knowledgeable, healthy, useful, cruelly, lawlessness, youtube, infallible, lousy, information, productive, wicked, activity, learn, competence, read, specific, fatass, eat, development, task, water, teach, misuse, carelessness, doctor, hobbies, body, drug, routine, gluttony, cold, vilify, addiction, laziest, area, exercise, gym, book, hobby, humanitarian, laziness, smoke, skill, tv, wake, snob, music, competent, personal, list, content, video, diet, weight, internet, bed, health, practice, game, language, sneakily, sport"]}

df_1 = pd.DataFrame(data=d)
df_2 = pd.DataFrame(data=d2)
with pd.option_context('display.max_colwidth', None):
    display(df_1)
    display(df_2)
